+++ op/strdel_test.go
@@ -13,8 +13,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 		name     string
 		path     []string
 		doc      interface{}
-		pos      int
-		length   int
+		pos      float64
+		length   float64
 		expected interface{}
 		oldValue interface{}
 		wantErr  bool
@@ -23,8 +23,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete in middle",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "hello, brave new world"},
-			pos:      5,
-			length:   12,
+			pos:      5.0,
+			length:   12.0,
 			expected: map[string]interface{}{"text": "helloworld"},
 			oldValue: "hello, brave new world",
 		},
@@ -32,8 +32,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete at start",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "hello world"},
-			pos:      0,
-			length:   6,
+			pos:      0.0,
+			length:   6.0,
 			expected: map[string]interface{}{"text": "world"},
 			oldValue: "hello world",
 		},
@@ -41,8 +41,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete at end",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "hello world"},
-			pos:      5,
-			length:   6,
+			pos:      5.0,
+			length:   6.0,
 			expected: map[string]interface{}{"text": "hello"},
 			oldValue: "hello world",
 		},
@@ -50,8 +50,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete unicode",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "你好，美丽的世界"},
-			pos:      2,
-			length:   4,
+			pos:      2.0,
+			length:   4.0,
 			expected: map[string]interface{}{"text": "你好世界"},
 			oldValue: "你好，美丽的世界",
 		},
@@ -59,8 +59,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete in nested",
 			path:     []string{"user", "bio"},
 			doc:      map[string]interface{}{"user": map[string]interface{}{"bio": "Golang dev"}},
-			pos:      2,
-			length:   4,
+			pos:      2.0,
+			length:   4.0,
 			expected: map[string]interface{}{"user": map[string]interface{}{"bio": "Go dev"}},
 			oldValue: "Golang dev",
 		},
@@ -68,8 +68,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete in array element",
 			path:     []string{"lines", "1"},
 			doc:      map[string]interface{}{"lines": []interface{}{"foo", "b-insert-ar", "baz"}},
-			pos:      1,
-			length:   8,
+			pos:      1.0,
+			length:   8.0,
 			expected: map[string]interface{}{"lines": []interface{}{"foo", "bar", "baz"}},
 			oldValue: "b-insert-ar",
 		},
@@ -77,8 +77,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete at root",
 			path:     []string{},
 			doc:      "a-bc",
-			pos:      1,
-			length:   1,
+			pos:      1.0,
+			length:   1.0,
 			expected: "abc",
 			oldValue: "a-bc",
 		},
@@ -86,32 +86,32 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:    "path not found",
 			path:    []string{"notfound"},
 			doc:     map[string]interface{}{"text": "abc"},
-			pos:     1,
-			length:  1,
+			pos:     1.0,
+			length:  1.0,
 			wantErr: true,
 		},
 		{
 			name:    "not a string",
 			path:    []string{"num"},
 			doc:     map[string]interface{}{"num": 123},
-			pos:     1,
-			length:  1,
+			pos:     1.0,
+			length:  1.0,
 			wantErr: true,
 		},
 		{
 			name:    "root not a string",
 			path:    []string{},
 			doc:     123,
-			pos:     1,
-			length:  1,
+			pos:     1.0,
+			length:  1.0,
 			wantErr: true,
 		},
 		{
 			name:     "delete position out of range",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      10,
-			length:   1,
+			pos:      10.0,
+			length:   1.0,
 			expected: map[string]interface{}{"text": "abc"},
 			oldValue: "abc",
 		},
@@ -119,8 +119,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete negative position",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      -1,
-			length:   1,
+			pos:      -1.0,
+			length:   1.0,
 			expected: map[string]interface{}{"text": "abc"},
 			oldValue: "abc",
 		},
@@ -128,8 +128,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete negative length",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      1,
-			length:   -1,
+			pos:      1.0,
+			length:   -1.0,
 			expected: map[string]interface{}{"text": "abc"},
 			oldValue: "abc",
 		},
@@ -137,8 +137,8 @@ func TestOpStrDel_Apply(t *testing.T) {
 			name:     "delete length out of range",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      1,
-			length:   10,
+			pos:      1.0,
+			length:   10.0,
 			expected: map[string]interface{}{"text": "a"},
 			oldValue: "abc",
 		},
@@ -165,19 +165,19 @@ func TestOpStrDel_Apply(t *testing.T) {
 }
 
 func TestOpStrDel_Op(t *testing.T) {
-	op := NewOpStrDelOperation([]string{"text"}, 1, 1)
+	op := NewOpStrDelOperation([]string{"text"}, 1.0, 1.0)
 	assert.Equal(t, internal.OpStrDelType, op.Op())
 }
 
 func TestOpStrDel_Code(t *testing.T) {
-	op := NewOpStrDelOperation([]string{"text"}, 1, 1)
+	op := NewOpStrDelOperation([]string{"text"}, 1.0, 1.0)
 	assert.Equal(t, internal.OpStrDelCode, op.Code())
 }
 
 func TestOpStrDel_NewOpStrDel(t *testing.T) {
 	path := []string{"user", "bio"}
-	pos := 2
-	length := 3
+	pos := 2.0
+	length := 3.0
 	op := NewOpStrDelOperation(path, pos, length)
 	assert.Equal(t, path, op.Path())
 	assert.Equal(t, pos, op.Pos)
