+++ op/strins_test.go
@@ -13,7 +13,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 		name     string
 		path     []string
 		doc      interface{}
-		pos      int
+		pos      float64
 		str      string
 		expected interface{}
 		oldValue interface{}
@@ -23,7 +23,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert in middle",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "hello world"},
-			pos:      5,
+			pos:      5.0,
 			str:      ", brave new",
 			expected: map[string]interface{}{"text": "hello, brave new world"},
 			oldValue: "hello world",
@@ -32,7 +32,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert at start",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "world"},
-			pos:      0,
+			pos:      0.0,
 			str:      "hello ",
 			expected: map[string]interface{}{"text": "hello world"},
 			oldValue: "world",
@@ -41,7 +41,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert at end",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "hello"},
-			pos:      5,
+			pos:      5.0,
 			str:      " world",
 			expected: map[string]interface{}{"text": "hello world"},
 			oldValue: "hello",
@@ -50,7 +50,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert unicode",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "你好世界"},
-			pos:      2,
+			pos:      2.0,
 			str:      "，美丽的",
 			expected: map[string]interface{}{"text": "你好，美丽的世界"},
 			oldValue: "你好世界",
@@ -59,7 +59,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert in nested",
 			path:     []string{"user", "bio"},
 			doc:      map[string]interface{}{"user": map[string]interface{}{"bio": "Go dev"}},
-			pos:      2,
+			pos:      2.0,
 			str:      "lang ",
 			expected: map[string]interface{}{"user": map[string]interface{}{"bio": "Golang  dev"}},
 			oldValue: "Go dev",
@@ -68,7 +68,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert in array element",
 			path:     []string{"lines", "1"},
 			doc:      map[string]interface{}{"lines": []interface{}{"foo", "bar", "baz"}},
-			pos:      1,
+			pos:      1.0,
 			str:      "-insert-",
 			expected: map[string]interface{}{"lines": []interface{}{"foo", "b-insert-ar", "baz"}},
 			oldValue: "bar",
@@ -77,7 +77,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert at root",
 			path:     []string{},
 			doc:      "abc",
-			pos:      1,
+			pos:      1.0,
 			str:      "-",
 			expected: "a-bc",
 			oldValue: "abc",
@@ -86,7 +86,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:    "path not found",
 			path:    []string{"notfound"},
 			doc:     map[string]interface{}{"text": "abc"},
-			pos:     1,
+			pos:     1.0,
 			str:     "-",
 			wantErr: true,
 		},
@@ -94,7 +94,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:    "not a string",
 			path:    []string{"num"},
 			doc:     map[string]interface{}{"num": 123},
-			pos:     1,
+			pos:     1.0,
 			str:     "-",
 			wantErr: true,
 		},
@@ -102,7 +102,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:    "root not a string",
 			path:    []string{},
 			doc:     123,
-			pos:     1,
+			pos:     1.0,
 			str:     "-",
 			wantErr: true,
 		},
@@ -110,7 +110,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert position out of range",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      10,
+			pos:      10.0,
 			str:      "X",
 			expected: map[string]interface{}{"text": "abcX"},
 			oldValue: "abc",
@@ -119,7 +119,7 @@ func TestOpStrIns_Apply(t *testing.T) {
 			name:     "insert negative position",
 			path:     []string{"text"},
 			doc:      map[string]interface{}{"text": "abc"},
-			pos:      -1,
+			pos:      -1.0,
 			str:      "X",
 			expected: map[string]interface{}{"text": "Xabc"},
 			oldValue: "abc",
@@ -147,18 +147,18 @@ func TestOpStrIns_Apply(t *testing.T) {
 }
 
 func TestOpStrIns_Op(t *testing.T) {
-	op := NewOpStrInsOperation([]string{"text"}, 1, "-")
+	op := NewOpStrInsOperation([]string{"text"}, 1.0, "-")
 	assert.Equal(t, internal.OpStrInsType, op.Op())
 }
 
 func TestOpStrIns_Code(t *testing.T) {
-	op := NewOpStrInsOperation([]string{"text"}, 1, "-")
+	op := NewOpStrInsOperation([]string{"text"}, 1.0, "-")
 	assert.Equal(t, internal.OpStrInsCode, op.Code())
 }
 
 func TestOpStrIns_NewOpStrInsOperation(t *testing.T) {
 	path := []string{"user", "bio"}
-	pos := 2
+	pos := 2.0
 	str := "abc"
 	op := NewOpStrInsOperation(path, pos, str)
 	assert.Equal(t, path, op.Path())
